Hotfix - Fix activity log comments (#464)
* Show comments on products

* Fixes to comments on activity log

* Apply fixes from StyleCI
