+++ packages/admin/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### [Unreleased]
+
+## Fixed
+
+- Fixed an issue where adding a comment to the activity log would error.
+- Comments should now show correctly on product editing pages.
+
 ### 2.0-beta14 - 2022-08-03
 
 ## Changed
